refactor(replay): move shared .ad vocabulary to its owner, packages/ad-script (#1555 review)
