+++ packages/ad-replay/src/index.ts
@@ -1,39 +1,34 @@
 /**
- * The `ad-replay` package façade (#1478 P5 stage D — narrowed).
+ * The `ad-replay` package façade (#1478 P5 stage D — narrowed; report-action/
+ * suggestion-ranking/vars/identity-vocabulary further narrowed by the P5
+ * review pass, "complete the binding façade instead of documenting
+ * deviations").
  *
  * The binding design (issue comment 5156017698) is `inspectAdReplay` +
  * `runAdReplay` and nothing else. Staged reality is wider than that: the
  * daemon wire-builders and handlers call several engine POLICY functions
- * directly rather than only through the two entrypoints (var substitution,
- * plan-digest, target-identity/-verification, report shaping). Every export
- * below is a `façade-deviation` from the binding-design ideal EXCEPT
- * `inspectAdReplay`/`runAdReplay` themselves; each deviation names its real
- * root consumer(s) so the PR body's deviation table is generated straight
- * from this file. No selector AST, no engine IR, no prepared-plan type, and
- * no internal subpath is exported — everything here has a live import site
- * outside this package (see `packages/ad-replay/src/index.ts`'s sibling
- * consumer grep in the P5 stage D commit message for the full map).
+ * directly rather than only through the two entrypoints (plan-digest,
+ * target-verification). Every export below is a `façade-deviation` from the
+ * binding-design ideal EXCEPT `inspectAdReplay`/`runAdReplay` themselves;
+ * each deviation names its real root consumer(s) so the PR body's deviation
+ * table is generated straight from this file. No selector AST, no engine IR,
+ * no prepared-plan type, and no internal subpath is exported — everything
+ * here has a live import site outside this package (see
+ * `packages/ad-replay/src/index.ts`'s sibling consumer grep in the P5 stage D
+ * commit message for the full map).
+ *
+ * `.ad` variable substitution (`vars.ts`), the local-identity + ancestry-
+ * prefix matching primitives and their diagnostic diffs (formerly part of
+ * `target-identity.ts`), and the divergence-report action shape/suggestion
+ * ranking (formerly `session-replay-report-action.ts` /
+ * `session-replay-suggestion-ranking.ts`) are NOT here: the review found
+ * these were never engine-owned policy reached only through the two
+ * entrypoints. Var substitution and identity matching are `.ad` script
+ * vocabulary the daemon and this engine both consume, so they moved to their
+ * proper shared owner, `@agent-device/ad-script`. Report-action/suggestion-
+ * ranking had no consumer inside this package at all — daemon-only — so they
+ * moved back to `src/daemon/handlers/`.
  */
-
-// ---------------------------------------------------------------------------
-// vars.ts — `.ad` variable substitution surface.
-// façade-deviation: the daemon resolves an action's variables directly
-// (`session-replay-target-verification.ts`, `session-replay-action-runtime.ts`)
-// and builds/collects env-derived scopes outside the step loop
-// (`session-replay-runtime.ts`, `session-replay-maestro-runtime.ts`,
-// `session-replay-runtime-failure.ts`) instead of only receiving resolved
-// actions back from `runAdReplay`.
-// ---------------------------------------------------------------------------
-export {
-  buildReplayVarScope,
-  collectReplayScrubbableVarValues,
-  collectReplayShellEnv,
-  parseReplayCliEnvEntries,
-  readReplayCliEnvEntries,
-  readReplayShellEnvSource,
-  resolveReplayAction,
-} from './internal/vars.ts';
-export type { ReplayVarScope } from './internal/vars.ts';
 
 // ---------------------------------------------------------------------------
 // plan-digest.ts — the `--from`/`--plan-digest` resume and save-script digest.
@@ -60,25 +55,20 @@ export { formatReplaySuccessMessage, runAdReplay } from './internal/step-loop.ts
 export type { AdReplayStepRuntime } from './internal/step-loop.ts';
 
 // ---------------------------------------------------------------------------
-// target-identity.ts — record/replay local-identity primitives (ADR 0012
-// decision 3).
-// façade-deviation: the record-time writer (`src/daemon/session-target-evidence.ts`)
-// and replay-time classification core (`session-replay-target-classification.ts`)
-// call these identity functions directly so both sides compute the SAME
-// ordinal by construction; `wait`'s landmark poll
-// (`src/commands/interaction/runtime/selector-wait.ts`) and the shared
-// replay-zone tree helpers (`src/replay/target-evidence-tree.ts`,
-// `src/replay/target-identity-node.ts`) also call in below `runAdReplay`.
+// target-identity.ts — record/replay-shared CLASSIFICATION core (ADR 0012
+// decision 3, replay-time verification paths 2-6).
+// façade-deviation: the daemon's replay-time classification core
+// (`session-replay-target-classification.ts`) and the record-time writer
+// (`src/daemon/session-target-evidence.ts`) call `classifyTargetBindingMatch`
+// directly so both sides compute the SAME verdict by construction, ahead of
+// and independent from any `runAdReplay` call. The local-identity +
+// ancestry-prefix matching primitives this module used to also export moved
+// to `@agent-device/ad-script` — they are `.ad` script vocabulary the
+// record-time writer, this classification core, and `wait`'s landmark poll
+// (`src/commands/interaction/runtime/selector-wait.ts`) all consume
+// directly, not engine policy (#1478 P5 review).
 // ---------------------------------------------------------------------------
-export {
-  annotationLocalIdentity,
-  classifyTargetBindingMatch,
-  firstAncestryMismatch,
-  identityFieldMismatches,
-  matchesAncestryPrefix,
-  matchesLocalIdentity,
-} from './internal/target-identity.ts';
-export type { LocalIdentity } from './internal/target-identity.ts';
+export { classifyTargetBindingMatch } from './internal/target-identity.ts';
 
 // ---------------------------------------------------------------------------
 // target-verification.ts — #1478 P5 stage C2a target-verification ENGINE
@@ -95,16 +85,6 @@ export {
   planPreDispatchTargetVerification,
 } from './internal/target-verification.ts';
 export type { ReplayPostDispatchMismatchEvidence } from './internal/target-verification.ts';
-
-// ---------------------------------------------------------------------------
-// session-replay-report-action.ts / session-replay-suggestion-ranking.ts —
-// the divergence-report action shape and repair-suggestion ranking.
-// façade-deviation: `session-replay-maestro-failure.ts`, `session-replay-heal.ts`,
-// and `session-replay-divergence.ts` build/rank divergence reports directly;
-// none of this rides back through `runAdReplay`'s return value.
-// ---------------------------------------------------------------------------
-export type { ReplayReportAction } from './internal/session-replay-report-action.ts';
-export { rankAndDedupeReplaySuggestions } from './internal/session-replay-suggestion-ranking.ts';
 
 // ---------------------------------------------------------------------------
 // selector-port.ts — the `ReplaySelectorPort` port TYPE only (#1478 P5 stage
