+++ packages/ad-replay/src/internal/target-identity.ts
@@ -1,68 +1,18 @@
 /**
  * ADR 0012 decision 3: the record/replay-shared CLASSIFICATION core over
- * versioned `.ad` target-binding evidence — local-identity + ancestry-prefix
- * matching, and `classifyTargetBindingMatch`'s replay-time verification
- * paths 2-6. Inert in migration step 3: nothing enforces parsed evidence at
- * replay time until step 4.
+ * versioned `.ad` target-binding evidence — `classifyTargetBindingMatch`'s
+ * replay-time verification paths 2-6. Inert in migration step 3: nothing
+ * enforces parsed evidence at replay time until step 4.
  *
  * The comment-line SERDE half (wire type, canonical field order,
  * normalization, size caps, payload parsing/validation) moved to
- * `@agent-device/ad-script` (#1478 P5 scoping dossier, "the codec seam") —
- * this module imports the shared types from there rather than declaring them.
+ * `@agent-device/ad-script` (#1478 P5 scoping dossier, "the codec seam").
+ * The local-identity + ancestry-prefix matching primitives and the bounded
+ * diagnostic diffs built on them also moved there (#1478 P5 review, "keep
+ * genuinely shared recording vocabulary in its proper shared owner") — this
+ * module imports them from there rather than declaring them, since decision
+ * 3's classification core is engine-owned policy, not script vocabulary.
  */
-
-import type { TargetAncestryEntry, TargetAnnotationV1 } from '@agent-device/contracts/replay';
-
-// ---------------------------------------------------------------------------
-// Local identity + ancestry-prefix matching (decision 3 "Local identity" /
-// "Ancestry"). Pure over the small structural shapes above — no tree
-// dependency, so both the writer (over `SnapshotNode`-derived values) and a
-// future replay verifier can share it verbatim.
-// ---------------------------------------------------------------------------
-
-export type LocalIdentity = { id?: string; role: string; label?: string };
-
-/** The recorded annotation's identity tier as a bare `LocalIdentity` (drop-empty-keys form). */
-export function annotationLocalIdentity(
-  recorded: Pick<TargetAnnotationV1, 'id' | 'role' | 'label'>,
-): LocalIdentity {
-  return {
-    ...(recorded.id !== undefined ? { id: recorded.id } : {}),
-    role: recorded.role,
-    ...(recorded.label !== undefined ? { label: recorded.label } : {}),
-  };
-}
-
-/**
- * Decision 3 "Local identity": id match wins outright when the recording
- * carries one ("a recorded id never matches a node without that id"); with
- * no recorded id, role+label must both match (label absent on both sides
- * counts as equal; present on exactly one side is a mismatch).
- */
-export function matchesLocalIdentity(candidate: LocalIdentity, recorded: LocalIdentity): boolean {
-  if (recorded.id !== undefined) return candidate.id === recorded.id;
-  return candidate.role === recorded.role && candidate.label === recorded.label;
-}
-
-/**
- * Decision 3 "Ancestry": leaf-anchored prefix match. `observed` must be at
- * least as long as `recorded`; each recorded entry's role must match exactly
- * and, when the recorded entry carries a label, so must the observed one (an
- * absent recorded label is unconstrained).
- */
-export function matchesAncestryPrefix(
-  observed: readonly TargetAncestryEntry[],
-  recorded: readonly TargetAncestryEntry[],
-): boolean {
-  if (observed.length < recorded.length) return false;
-  for (const [index, entry] of recorded.entries()) {
-    const candidate = observed[index];
-    if (!candidate) return false;
-    if (candidate.role !== entry.role) return false;
-    if (entry.label !== undefined && candidate.label !== entry.label) return false;
-  }
-  return true;
-}
 
 // ---------------------------------------------------------------------------
 // Classification core (decision 3 "Replay-time verification", paths 2-6;
@@ -149,59 +99,4 @@ export function classifyTargetBindingMatch(
       : { path: 6, outcome: 'unverifiable', reason: 'signal-isolated-wrong' };
   }
   return { path: 6, outcome: 'unverifiable', reason: 'no-signal-isolation' };
-}
-
-// ---------------------------------------------------------------------------
-// Diagnostic diffs (decision 3): bounded, best-effort mismatch descriptions
-// shared by the record-time classification core and replay-time verification
-// (#1478 P5 stage C2a) — moved here verbatim from
-// `src/daemon/handlers/session-replay-target-classification.ts` so both
-// callers depend on one definition instead of two copies.
-// ---------------------------------------------------------------------------
-
-export function identityFieldMismatches(
-  recorded: TargetAnnotationV1,
-  observed: LocalIdentity,
-): string[] {
-  const mismatches: string[] = [];
-  if (recorded.id !== observed.id) {
-    mismatches.push(`id: recorded=${recorded.id ?? '(none)'} observed=${observed.id ?? '(none)'}`);
-  }
-  if (recorded.role !== observed.role) {
-    mismatches.push(`role: recorded=${recorded.role} observed=${observed.role}`);
-  }
-  if (recorded.label !== observed.label) {
-    mismatches.push(
-      `label: recorded=${recorded.label ?? '(none)'} observed=${observed.label ?? '(none)'}`,
-    );
-  }
-  return mismatches;
-}
-
-function describeAncestryEntry(entry: TargetAncestryEntry | undefined): string {
-  return entry ? `${entry.role}${entry.label ? `/${entry.label}` : ''}` : '(missing)';
-}
-
-function ancestryEntryMismatches(
-  expected: TargetAncestryEntry,
-  actual: TargetAncestryEntry | undefined,
-): boolean {
-  if (!actual) return true;
-  if (actual.role !== expected.role) return true;
-  return expected.label !== undefined && actual.label !== expected.label;
-}
-
-/** Leaf-anchored prefix: the first divergence explains everything after it. */
-export function firstAncestryMismatch(
-  recordedAncestry: readonly TargetAncestryEntry[],
-  observedAncestry: readonly TargetAncestryEntry[],
-): string[] {
-  for (const [index, expected] of recordedAncestry.entries()) {
-    const actual = observedAncestry[index];
-    if (!ancestryEntryMismatches(expected, actual)) continue;
-    return [
-      `ancestry[${index}]: recorded=${describeAncestryEntry(expected)} observed=${describeAncestryEntry(actual)}`,
-    ];
-  }
-  return [];
 }
