+++ packages/ad-script/src/internal/__tests__/vars.test.ts
@@ -8,7 +8,7 @@ import {
   resolveReplayAction,
   resolveReplayString,
 } from '../vars.ts';
-import { parseReplayScriptDetailed, readReplayScriptMetadata } from '@agent-device/ad-script';
+import { parseReplayScriptDetailed, readReplayScriptMetadata } from '../script.ts';
 import type { SessionAction } from '@agent-device/contracts/session';
 
 const LOC = { file: 'test.ad', line: 1 };
