+++ src/daemon/handlers/session-replay-maestro-failure.ts
@@ -12,7 +12,8 @@ import { formatScriptArg } from '@agent-device/ad-script';
 import { getRequestSignal } from '../../request/cancel.ts';
 import { SessionStore } from '../session-store.ts';
 import type { DaemonRequest, DaemonResponse, SessionState } from '../types.ts';
-import { rankAndDedupeReplaySuggestions, type ReplayReportAction } from '@agent-device/ad-replay';
+import type { ReplayReportAction } from './session-replay-report-action.ts';
+import { rankAndDedupeReplaySuggestions } from './session-replay-suggestion-ranking.ts';
 import {
   buildReplayDivergenceSuggestionForNode,
   buildDivergenceScreen,
