+++ src/daemon/handlers/session-replay-report-action.ts
